+++ example/lib/main.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+import 'src/minimal_app.dart';
+
+void main() async {
+  runApp(const MinimalApp());
+}