+++ test/src/mm_notifier_test.dart
@@ -1,0 +1,78 @@
+// Let's make the tests more readable and explicit
+// ignore_for_file: cascade_invocations
+import 'package:flutter_test/flutter_test.dart';
+
+import 'utils/test_common.dart';
+
+void main() {
+  group('MMNotifier', () {
+    test('notifies listeners upon state changes', () {
+      var notified = 0;
+      void l() => notified++;
+
+      final notifier = TNotifier();
+      notifier.addListener(l);
+      expect(notified, 0);
+      expect(notifier.state.value, 0);
+
+      notifier.increment();
+      expect(notified, 1);
+      expect(notifier.state.value, 1);
+
+      notifier.increment();
+      expect(notified, 2);
+      expect(notifier.state.value, 2);
+    });
+
+    test('stops notifying removed listeners', () {
+      var notified = 0;
+      void l() => notified++;
+
+      final notifier = TNotifier();
+      notifier.addListener(l);
+      expect(notified, 0);
+      expect(notifier.state.value, 0);
+
+      notifier.increment();
+      expect(notified, 1);
+      expect(notifier.state.value, 1);
+
+      notifier.removeListener(l);
+      notifier.increment();
+      expect(notified, 1);
+      expect(notifier.state.value, 2);
+    });
+
+    test('selects part of state', () {
+      var stateNotified = 0;
+      void stateL() => stateNotified++;
+
+      var selectionNotified = 0;
+      void selectionL() => selectionNotified++;
+
+      final notifier = TNotifier();
+      notifier.addListener(stateL);
+
+      final selected = notifier.select((final state) => state.text);
+      selected.addListener(selectionL);
+
+      notifier.increment();
+      expect(notifier.state.value, 1);
+      expect(notifier.state.text, '');
+      expect(stateNotified, 1);
+      expect(selectionNotified, 0);
+
+      notifier.increment();
+      expect(notifier.state.value, 2);
+      expect(notifier.state.text, '');
+      expect(stateNotified, 2);
+      expect(selectionNotified, 0);
+
+      notifier.append('mutated');
+      expect(notifier.state.value, 2);
+      expect(notifier.state.text, 'mutated');
+      expect(stateNotified, 3);
+      expect(selectionNotified, 1);
+    });
+  });
+}